Resolve project structure refactor #5
