+++ combat/emotions.py
