+++ start/start_screen.py
@@ -1,7 +1,7 @@
 # todo get the user to select a main task to do while playing the game
  
 import pygame
-from Helper.button import Button
+from start.button import Button
 
 class TaskTree:
     """
@@ -15,7 +15,7 @@ class TaskTree:
         self.display.fill("red")
 
         # add a button
-        start_img = pygame.image.load("Assets/start_button.png").convert_alpha()
+        start_img = pygame.image.load("start/start_button.png").convert_alpha()
         start_button = Button(self.display, 100, 100, start_img, 0.4)
         start_button.draw()
